% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy.default}
\alias{accuracy.default}
\title{Generic method for accuracy}
\usage{
\method{accuracy}{default}(f, ...)
}
\arguments{
\item{f}{an object of class forecast, or a numerical vector containing forecasts.
It will also work with Arima, ets and lm objects if x is omitted - in which case
in-sample accuracy measures are returned.}

\item{...}{other arguments (ignored)}
}
\description{
Generic method for accuracy
}
\seealso{
\code{\link[forecast]{accuracy}}, \code{\link{accuracy.hybridModel}}
}

