% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy.hybridModel}
\alias{accuracy.hybridModel}
\title{Accuracy measures for hybridModel objects}
\usage{
accuracy.hybridModel(x)
}
\arguments{
\item{x}{The input hybridModel
#' @seealso \code{\link{accuracy}}}
}
\description{
Return the in-sample accuracy measures for the component models of the hybridModel
}

