% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy}
\alias{accuracy}
\title{Generic method for accuracy}
\usage{
accuracy(f, ...)
}
\arguments{
\item{f}{The input object}

\item{...}{Other arguments (ignored).}
}
\description{
Generic method for accuracy
}

