% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvts.R
\name{checkCVArguments}
\alias{checkCVArguments}
\title{Validate that CV window parameters are valid}
\usage{
checkCVArguments(x, windowSize, maxHorizon)
}
\arguments{
\item{x}{the input time series.}

\item{windowSize}{length of the window to build each model. When \code{rolling == FALSE},
the each model will be fit to a time series of this length, and when \code{rolling == TRUE}
the first model will be fit to a series of this length and grow by one each iteration.}

\item{maxHorizon}{maximum length of the forecast horizon to use for computing errors.}
}
\description{
Validate that CV window parameters are valid
}
