% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkModelArgs}
\alias{checkModelArgs}
\title{Helper function to test all the model arguments (e.g. a.args, e.args, etc)}
\usage{
checkModelArgs(modelArguments, models)
}
\arguments{
\item{modelArguments}{A list of containing the model arguments}

\item{models}{A character vector containing all the model codes}
}
\description{
Helper function to test all the model arguments (e.g. a.args, e.args, etc)
}
