% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkParallelArguments}
\alias{checkParallelArguments}
\title{Helper function to check the that the parallel arguments are valid}
\usage{
checkParallelArguments(parallel, num.cores)
}
\arguments{
\item{parallel}{A logic to indicate if paralle processing should be used}

\item{num.cores}{An integer for the number of threads to use}
}
\description{
Helper function to check the that the parallel arguments are valid
}
