% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvts.R
\name{cvts}
\alias{cvts}
\title{Cross validation for time series}
\usage{
cvts(x, FUN = NULL, FCFUN = NULL, rolling = FALSE, windowSize = 84,
  useHorizon = 5, maxHorizon = 5, horizonAverage = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{the input time series}

\item{FUN}{the model function used. See details.}

\item{FCFUN}{a function that proces point forecasts for the model function. This defaults to \code{\link{forecast()}}.
See details.}

\item{rolling}{should a rolling procedure be used? If TRUE, nonoverlapping windows of size \code{maxHorizon}
will be used for fitting each model. If FALSE, the size of the dataset used for training will grow
by one each iteration.}

\item{maxHorizon}{maximum length of the forecast horizon to use for computing errors}

\item{horizonAverage}{should the final errors be an average over all forecast horizons up to \code{maxHorizon} instead of producing
metrics for each individual horizon?}

\item{verbose}{should the current progress be printed to the console?}

\item{windoSize}{length of the window to build each model. When \code{rolling == TRUE}, the each model will be
fit to a time series of this length, and when \code{rolling == FALSE} the first model will be fit to a series
of this length and grow by one each iteration}
}
\description{
Perform cross validation on time series
}
\details{
Cross validation of time series is more complicated than regular k-folds of leave one out cross validation of datasets
without serial correlation since observations \eqn{x_t}{x[t]} and \eqn{x_{t+n}}{x[t+n]} are not independent. The \code{cvts()} function overcomes
this obstacle using two methods: rolling cross validation where an initial training window is used along with a forecast horizon
and the initial window used for training grows by one each round until the training window and the forecast horizon capture the
entire series or a nonrolling approach where a fixed training horizon is used that is shifted forward by the forecast horizon
after each iteration.
\cr
\cr
For the rolling approach, training points are heavily recycled, both in terms of used for fitting
and in generating forecast errors at each of the forecast horizons from \code{1:maxHorizon}. In constrast, the models fit with
the nonrolling approach share less overlap, and the predicted forecast values are also only compared to the actual values once.
former approach is similar to leave one out cross validatoin whlie the latter resembles k-fold cross validation. As a result,
rolling cross validation can use far more iterations and computationally take longer to complete, but a disadvantage of the
nonrolling approach is the greater variance and general instability of cross validated errors.
}
\examples{

cvmod1 <- cvts(AirPassengers, FUN = ets, FCFUN = forecast, rolling = TRUE, windowSize = 48, maxHorizon = 12)
cvmod2 <- cvts(AirPassengers, FUN = hybridModel, FCFUN = forecast, rolling = TRUE, windowSize = 48, maxHorizon = 12)
}


