% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{fitted.hybridModel}
\alias{fitted.hybridModel}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{hybridModel}(object, individual = FALSE, ...)
}
\arguments{
\item{object}{The input hybridModel}

\item{individual}{If \code{TRUE}, return the fitted values of the component models instead
of the fitted values for the whole ensemble model.}

\item{...}{Other arguments (ignored).}
}
\value{
The fitted values of the ensemble or individual component models
}
\description{
Extract the model fitted values from the \code{hybridModel} object.
}
\seealso{
\code{\link{accuracy}}
}

