% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.hybridModel.R
\name{forecast.hybridModel}
\alias{forecast.hybridModel}
\title{Hybrid forecast}
\usage{
\method{forecast}{hybridModel}(object, h = ifelse(object$frequency > 1, 2 *
  object$frequency, 10))
}
\arguments{
\item{object}{A hybrid time series model fit with hybridModel()}

\item{h}{number of periods to forecast ahead}
}
\description{
Forecast method for hybrid models
}
\details{
more detailed description here
}
\examples{
print("hello world")

}
\seealso{
\code{\link{hybridModel}}
}

