% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getModel}
\alias{getModel}
\title{Return a forecast model function for a given model character}
\usage{
getModel(modelCharacter)
}
\arguments{
\item{modelCharacter}{a single character representing one of the models from the \code{models}
argument passed to \link{hybridModel}}
}
\description{
Convert the single-letter representation used in the "forecastHybrid" package to the
corresponding model function from the "forecast" package
}
\examples{
forecastHybrid:::getModel("a")
forecastHybrid:::getModel("s")
forecastHybrid:::getModel("z")
}
\seealso{
\code{\link{hybridModel}}
}
