% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getModelName}
\alias{getModelName}
\title{Translate character to model name}
\usage{
getModelName(modelCharacter)
}
\arguments{
\item{modelCharacter}{a single character representing one of the models from the \code{models}
argument passed to \link{hybridModel}}
}
\description{
Convert the single-letter representation used in the "forecastHybrid" package to the
corresponding function name from the "forecast" package
}
\examples{
forecastHybrid:::getModelName("a")
forecastHybrid:::getModelName("s")
forecastHybrid:::getModelName("z")
}
\seealso{
\code{\link{hybridModel}}
}
