% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{hybridModel}
\alias{hybridModel}
\title{Hybrid time series modelling}
\usage{
hybridModel(y, models = "aenst", lambda = NULL, a.args = NULL,
  e.args = NULL, n.args = NULL, s.args = NULL, t.args = NULL,
  weights = c("equal", "insample.errors", "cv.errors"),
  errorMethod = c("RMSE", "MAE", "MASE"), parallel = TRUE, num.cores = 2L)
}
\arguments{
\item{y}{a numeric vector or time series}

\item{models}{a character string of up to five characters indicating which contributing models to use: 
a (\code{auto.arima()}), e (\code{ets()}), n (\code{nnetar()}), s (\code{stlm()}) and t (\code{tbats()})}

\item{lambda}{Box-Cox transformation parameter.  Ignored if NULL.  Otherwise, data transformed before model is estimated.}

\item{a.args}{Arguments to pass to \code{auto.arima()}}

\item{e.args}{Arguments to pass to \code{ets()}}

\item{n.args}{Arguments to pass to \code{nnetar()}}

\item{s.args}{Arguments to pass to \code{stlm()}}

\item{t.args}{Arguments to pass to \code{tbats()}}

\item{weights}{Method for weighting the forecasts of the various contributing
models.  Defaults to equal, which has shown to be robust and surprisingly better
in many cases than giving more weight to models with better past performance.}

\item{errorMethod}{Method of measuring accuracy to use if weights are not 
to be equal.}

\item{parallel}{Should parallel processing be used?}

\item{num.cores}{If \code{parallel=TRUE}, how many cores to use.}
}
\value{
An object of class hybridModel
}
\description{
Create a hybrid time series model from one to four contributing models
}
\details{
more detailed description here
}
\examples{
mod <- hybridModel(AirPassengers)
plot(forecast(mod))

}
\seealso{
\code{\link{forecast.hybridModel}}
}

