% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.hybridModel}
\alias{is.hybridModel}
\title{Test if the object is a hybridModel object}
\usage{
is.hybridModel(x)
}
\arguments{
\item{x}{the input object.}
}
\value{
A boolean indicating if the object is a \code{hybridModel} is returned.
}
\description{
Test if the object is a \code{hybridModel} object.
}
