% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{plot.thetam}
\alias{plot.thetam}
\title{Plot components from Theta model}
\usage{
\method{plot}{thetam}(x, ...)
}
\arguments{
\item{x}{Object of class "thetam".}

\item{...}{Other plotting parameters passed through to \code{plot}}
}
\value{
None.  Function produces a plot.
}
\description{
Produces a plot of the level components from the ETS model underlying a Theta model
}
\author{
Peter Ellis
}

