% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plotFitted}
\alias{plotFitted}
\title{Plot the fitted values of a hybridModel object}
\usage{
plotFitted(x, ggplot, ...)
}
\arguments{
\item{x}{an object of class hybridModel to plot.}

\item{ggplot}{should the \code{\link{autoplot}} function be used (when available) for the plots?}

\item{...}{other arguments passed to \link{plot}.}
}
\description{
Plot a fitted values of the hybridModel.
}
