% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prepareTimeseries}
\alias{prepareTimeseries}
\title{Helper function to validate and clean the input timeseries}
\usage{
prepareTimeseries(y)
}
\arguments{
\item{y}{The input timeseries}
}
\description{
Helper function to validate and clean the input timeseries
}
