% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{removeModels}
\alias{removeModels}
\title{Helper function to remove models that require more data}
\usage{
removeModels(y, models)
}
\arguments{
\item{y}{The input time series}

\item{models}{The model codes to test}
}
\description{
Helper function to remove models that require more data
}
