% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{residuals.hybridModel}
\alias{residuals.hybridModel}
\title{Extract Model Residuals}
\usage{
\method{residuals}{hybridModel}(object, individual = FALSE, ...)
}
\arguments{
\item{object}{The input hybridModel}

\item{individual}{If \code{TRUE}, return the residuals of the component models instead
of the residuals for the whole ensemble model.}

\item{...}{Other arguments (ignored).}
}
\value{
The residuals of the ensemble or individual component models
}
\description{
Extract the model residuals from the \code{hybridModel} object.
}
\seealso{
\code{\link{accuracy}}
}

