% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.hybridModel}
\alias{summary.hybridModel}
\title{Print a summary of the hybridModel object}
\usage{
\method{summary}{hybridModel}(x)
}
\arguments{
\item{x}{the input \code{hybridModel} object.}
}
\description{
Print a summary of the hybridModel object
}
\details{
Print the names of the individual component models and their weights.
}
