% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiefModel.R
\name{thiefModel}
\alias{thiefModel}
\title{Forecast ensemble using THieF}
\usage{
thiefModel(y, models = "aefnt", h = 2 * frequency(x), verbose = TRUE)
}
\description{
Create a forecast ensemble using the theif() model
}
\details{
Use the "thief" package method for reconciling forecasts across the temporal hierarchy.
The base models to be included in the ensemble are the same as those in \text{hybridModel}, but
the \code{stlm} model cannot be included since it requires seasonal data.
}
\seealso{
\code{\link{[thief]thief}}

\code{\link{hybridModel}}
}
\author{
David Shaub
}
