% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tsSubsetWithIndices}
\alias{tsSubsetWithIndices}
\title{Subset time series with provided indices}
\usage{
tsSubsetWithIndices(x, indices)
}
\arguments{
\item{x}{A time series object}

\item{indices}{A contiguous vector of indices to use for subsetting}
}
\value{
A time series object appropriately subsetted using provided indices
}
\description{
Use provided indices to subset a time series. The provided indices must be contiguous
}
\examples{
tsSubsetWithIndices(AirPassengers, c(3:10))
}
\author{
Ganesh Krishnan
}
