% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{unwrapParallelModels}
\alias{unwrapParallelModels}
\title{Helper function used to unpack the fitted model objects from a list}
\usage{
unwrapParallelModels(fitModels, expandedModels)
}
\arguments{
\item{fitModels}{A list containing the models to include in the ensemble}

\item{expandedModels}{A character vector from the \code{models} argument of \link{hybridModel}}
}
\description{
Helper function used to unpack the fitted model objects from a list
}
\details{
See usage inside the \code{hybridModel} function.
}
\seealso{
\code{\link{hybridModel}}
}
