% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy.cvts}
\alias{accuracy.cvts}
\title{Accuracy measures for cross-validated time series}
\usage{
\method{accuracy}{cvts}(f, ...)
}
\arguments{
\item{f}{a \code{cvts} objected created by \code{\link{cvts}}.}

\item{...}{other arguments (ignored).}
}
\description{
Returns range of summary measures of the cross-validated forecast accuracy
for \code{cvts} objects.
}
\details{
Currently the method only implements \code{ME}, \code{RMSE}, and \code{MAE}. The accuracy measures
\code{MPE}, \code{MAPE}, and \code{MASE} are not calculated. The accuracy is calculated for each
forecast horizon up to \code{maxHorizon}
}

