% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{accuracy.hybridModel}
\alias{accuracy.hybridModel}
\title{Accuracy measures for hybridModel objects}
\usage{
\method{accuracy}{hybridModel}(object, individual = FALSE, ..., f = NULL)
}
\arguments{
\item{object}{the input hybridModel.}

\item{individual}{if \code{TRUE}, return the accuracy of the component models instead
of the accuracy for the whole ensemble model.}

\item{...}{other arguments (ignored).}

\item{f}{Deprecated. Please use `object` instead.}
}
\value{
The accuracy of the ensemble or individual component models.
}
\description{
Accuracy measures for hybridModel
objects.
}
\details{
Return the in-sample accuracy measures for the component models of the hybridModel
}
\seealso{
\code{\link[forecast]{accuracy}}
}
\author{
David Shaub
}
