% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy.hybridModel}
\alias{accuracy.hybridModel}
\title{Accuracy measures for hybridModel objects}
\usage{
\method{accuracy}{hybridModel}(f, individual = FALSE)
}
\arguments{
\item{f}{The input hybridModel}

\item{individual}{If \code{TRUE}, return the accuracy of the component models instead
of the accuracy for the whole ensemble model.}
}
\value{
The accuracy of the ensemble or individual component models
}
\description{
Return the in-sample accuracy measures for the component models of the hybridModel
}
\seealso{
\code{\link{accuracy}}
}

