% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{accuracy}
\alias{accuracy}
\title{Generic method for accuracy}
\usage{
accuracy(f, ...)
}
\arguments{
\item{f}{the input object.}

\item{...}{other arguments (ignored).}
}
\description{
Generic method for accuracy
}
\seealso{
\code{\link[forecast]{accuracy}}, \code{\link{accuracy.hybridModel}}
}

