% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{hybridModel}
\alias{hybridModel}
\title{Hybrid time series modelling}
\usage{
hybridModel(y, models = "aten", xreg = NULL, weights = c("equal",
  "insample.errors", "cv.errors"), errorMethod = c("rmse", "mae", "mase"),
  parallel = TRUE, num.cores = 2)
}
\arguments{
\item{y}{a numeric vector or time series}

\item{models}{a four character string indicating which contributing models to use: 
a (\code{auto.arima()}), t (\code{tbats()}), e (\code{ets()}) and n (\code{nnetar()})}

\item{xreg}{Optionally, a vector or matrix of external regressors, which must 
have the same number of rows as y.  Only used if 'a' is part of the models string
ie if \code{auto.arima} is one of the contributing models}

\item{weights}{Method for weighting the forecasts of the various contributing
models.  Defaults to equal, which has shown to be robust and surprisingly better
in many cases than giving more weight to models with better past performance.}

\item{parallel}{Should parallel processing be used?}

\item{num.cores}{If \code{parallel=TRUE}, how many cores to use.}

\item{errorMethod.}{Method of measuring accuracy to use if weights are not 
to be equal.}
}
\description{
Create a hybrid time series model from one to four contributing models
}
\details{
more detailed description here
}
\examples{
print("hello world")

}
\seealso{
\code{\link{forecast.hybridModel}}
}

