% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{plot.thetam}
\alias{plot.thetam}
\title{Plot components from Theta model}
\usage{
plot.thetam(object, ...)
}
\arguments{
\item{object}{Object of class "thetam".}
}
\value{
None.  Function produces a plot.
}
\description{
Produces a plot of the level components from the ETS model underlying a Theta model
}
\author{
Peter Ellis
}

