% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModel.R
\name{print.hybridModel}
\alias{print.hybridModel}
\title{Print information about the hybridModel object}
\usage{
\method{print}{hybridModel}(x, ...)
}
\arguments{
\item{x}{the input \code{hybridModel} object.}

\item{...}{other arguments (ignored).}
}
\description{
Print information about the \code{hybridModel} object.
}
\details{
Print the names of the individual component models and their weights.
}

